+++ components/shared/Card/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Card';
+export * from './Card';
